+++ src/interp1d/strategies/linear.rs
@@ -5,7 +5,7 @@ use num_traits::{Num, NumCast};
 
 use crate::{interp1d::Interp1D, BuilderError, InterpolateError};
 
-use super::{Strategy, StrategyBuilder};
+use super::{Interp1DStrategy, Interp1DStrategyBuilder};
 
 /// Linear Interpolation Strategy
 #[derive(Debug)]
@@ -13,7 +13,7 @@ pub struct Linear {
     pub extrapolate: bool,
 }
 
-impl<Sd, Sx, D> StrategyBuilder<Sd, Sx, D> for Linear
+impl<Sd, Sx, D> Interp1DStrategyBuilder<Sd, Sx, D> for Linear
 where
     Sd: Data,
     Sd::Elem: Num + PartialOrd + NumCast + Copy + Debug + Sub,
@@ -34,7 +34,7 @@ where
     }
 }
 
-impl<Sd, Sx, D> Strategy<Sd, Sx, D> for Linear
+impl<Sd, Sx, D> Interp1DStrategy<Sd, Sx, D> for Linear
 where
     Sd: Data,
     Sd::Elem: Num + PartialOrd + NumCast + Copy + Debug + Sub,
@@ -64,8 +64,20 @@ where
 
         // do interpolation
         Zip::from(target).and(y1).and(y2).for_each(|t, &y1, &y2| {
-            *t = Interp1D::<Sd, Sx, D, Self>::calc_frac((x1, y1), (x2, y2), x);
+            *t = Self::calc_frac((x1, y1), (x2, y2), x);
         });
         Ok(())
+    }
+}
+
+impl Linear {
+    /// linearly interpolate/exrapolate between two points
+    pub(crate) fn calc_frac<T>((x1, y1): (T, T), (x2, y2): (T, T), x: T) -> T
+    where
+        T: Num + Copy,
+    {
+        let b = y1;
+        let m = (y2 - y1) / (x2 - x1);
+        m * (x - x1) + b
     }
 }